docs: Name 구현 및 Member 에서 이동
